Move types,funcs in separate files for convenience
